GoogleService/requirements.txt
google_auth_oauthlib==0.4.4
oauth2client==4.1.3
gspread==3.1.0
google_api_python_client==2.8.0
protobuf==3.17.3
